requirements.txt
torch
torchvision
torchaudio
hydra-core
einops
numpy
omegaconf